refactor(tui): extract projects CRUD + repo discovery into methods_projects (server.py god-file slice R5)
